@@ -1,6 +1,7 @@
 {
   "name": "functions",
   "description": "Cloud Functions for Firebase",
+  "type": "module",
   "scripts": {
     "serve": "NODE_ENV=development firebase emulators:start --only functions",
     "shell": "NODE_ENV=development firebase functions:shell",
@@ -10,14 +11,13 @@
     "runtime": "firebase functions:config:get > .runtimeconfig.json"
   },
   "engines": {
-    "node": "16"
+    "node": "20"
   },
   "main": "index.js",
   "dependencies": {
     "firebase-admin": "^12.0.0",
     "firebase-functions": "^4.6.0"
   },
-  "devDependencies": {
-  },
+  "devDependencies": {},
   "private": true
 }
